Correct gazebo-model file placement

+++ install.bash
@@ -7,6 +7,7 @@ ROS_VER=kinetic
 [ "$UBUNTU_VER" = "trusty" ] && ROS_VER=indigo
 
 sudo apt-get install -y ros-${ROS_VER}-desktop-full
+sudo apt-get install -y mktemp
 
 source ~/catkin_ws/devel/setup.bash || { echo "catkin_ws is not working correctly"; exit 1; }
 
@@ -29,14 +30,17 @@ roscd && cd .. && source ~/catkin_ws/devel/setup.bash && catkin_make
 source ~/catkin_ws/devel/setup.bash
 
 mkdir -p ~/.gazebo/models && cd ~/.gazebo/models
-[ -e "ground_plane" ] || \
+if [ ! -e "ground_plane" ]; then
 	cd /tmp && \
+	TMPDIR=$(mktemp -d tmp.XXXXXXXXXX) && \
+	cd $TMPDIR && \
 	wget -l1 -np -nc -r "http://models.gazebosim.org/sun" --accept=gz && \
 	wget -l1 -np -nc -r "http://models.gazebosim.org/ground_plane" --accept=gz && \
 	wget -l1 -np -nc -r "http://models.gazebosim.org/gas_station" --accept=gz && \
 	cd "models.gazebosim.org" && \
 	for i in *; do tar -zvxf "$i/model.tar.gz"; done && \
 	cp -vfR * ~/.gazebo/models/
+fi
 
 ###HOW TO VERIFY###
 # roslaunch raspimouse_gazebo raspimouse_with_samplemaze.launch
